Refactor: Decouple Gamification Logic from Portfolio Service
